aarch64: Adjust aarch64/spaceship_1.C testcase for recent changes [PR121732]
In r16-3414 libstdc++ changed ABI for (still experimental C++20) and uses
unordered value -128 instead of 2.  Generally the change improved code
generation on all targets tested, see
https://gcc.gnu.org/pipermail/gcc-patches/2025-August/693534.html
for details.
In r16-3474 I've adjusted the middle-end and backends to use that value.
This apparently broke the spaceship_1.C test on aarch64 which scans the
exact function bodies which are now different.

The following patch adjusts the full body patterns to match.  On these
2 routines, the generated code is 1 insn longer than in the past, so if
you have ideas how to change the code generation for the common case of
-1, 0, 1, -128 value, maybe it could be improved.

2025-09-04  Jakub Jelinek  <jakub@redhat.com>

	PR testsuite/121732
	PR target/117013
	* g++.target/aarch64/spaceship_1.C: Adjust expected fn bodies
	for _Z8ss_floatff and _Z9ss_doubledd.