c: Allow translations of a lot of C FE messages
I've noticed a lot of diagnostic messages in the C FE aren't marked
for translation.
The reason is some weird coding style which wraps the string
literals into (), especially when they don't fit on a single line.
With that fixed, there were 83 unique similar messages
"both %<something%> and %<something%> in declaration specifiers"
marked for translation, which is very unfriendly to translators,
the patch brings that down to 4 (if it was ok to change order,
it could be even 3):
msgid "both %qs and %qs in declaration specifiers"
msgid "both %qs and %<__int%d%> in declaration specifiers"
msgid "both %qs and %<_Float%d%s%> in declaration specifiers"
msgid "both %<__int%d%> and %qs in declaration specifiers"

2025-09-05  Jakub Jelinek  <jakub@redhat.com>

	* c-decl.cc (pushtag): Remove ()s around string literal
	in call to diagnostic function.
	(diagnose_mismatched_decls): Likewise.
	(c_check_switch_jump_warnings): Likewise.
	(grokdeclarator): Likewise.
	(warn_cxx_compat_finish_struct): Likewise.
	(build_enumerator): Formatting fix.
	(declspecs_add_type): Remove ()s around string literal
	in call to diagnostic function, simplify
	"both %<something%> and %<something%>" starting format
	strings to "both %qs and %qs" with appropriate arguments.
	Formatting fixes.
	* c-typeck.cc (build_external_ref): Remove ()s around string
	literal in call to diagnostic function.
	(build_conditional_expr): Likewise.
	* c-parser.cc (c_parser_transaction): Use G_() around string
	literals.  Formatting fix.
	(c_parser_transaction_expression): Likewise.