phiopt: Improve locations for factor out conditional operation [PR108466]
This improves the locations for the phi args and the newly created statement.
Since this is a factorization/commonizing in one case the location for the new
statement will not always be set correctly either way.
The new locations on the new phi will either be the old location of the argument
to the phi or the location of the defining statement (if it exists).
The new statement will be either the location of the phi or the location
of the defining statements if the location of the phi is unknown.

This fixes the location of an uninitialized variable warning too.

Bootstrapped and tested on x86_64-linux-gnu.

	PR tree-optimization/108466
gcc/ChangeLog:

	* tree-ssa-phiopt.cc (factor_out_conditional_operation): Give better
	locations to the new phi args and the new statement.

gcc/testsuite/ChangeLog:

	* gcc.dg/uninit-pr108466-1.c: New test.

Signed-off-by: Andrew Pinski <andrew.pinski@oss.qualcomm.com>