[RISC-V] Fix ordering of pipeline models
I missed that the new ascalon pipeline description was put into the wrong place
during review.  The net is tests which wanted to use generic-ooo explicitly for
stability in the test output ended up getting  a different pipeline model and
different codegen than the test expected.

This tripped a small number of vsetvl failures in the testsuite.

This has spun on riscv64-elf and riscv32-elf in my tester and fixes the
regression.  I'm going to go ahead and push it as I'm likely offline this
afternoon/evening and don't want anyone else to waste their time chasing the
regression down.

gcc/
	* config/riscv/riscv-opts.h (riscv_microarchitecture_type): Fix ordering.