ada: New preprocessing option for emitting empty comments for deleted lines
When integrated preprocessing is done on a source file with lines that are
filtered out, by default this can result in multiple blank lines being emitted,
but this can clash with the style switch -gnatyu, which will flag cases of
multiple blank lines.  A new preprocessing option "-e" is added to allow
outputting "empty" comment lines rather than blank lines (the comments consist
of "--!" and no other text).  This option is also available for the gnatprep
tool.  This behavior is the default when integrated preprocessing is done
without generating a ".prep" output file with -gnateG, but when -gnateG is
used, blank lines are still substituted by default for directives and removed
source lines (for compatibility with long-existing behavior).  The -gnateG
switch is also extended to allow appending 'b', 'c', or 'e' at the end of
the switch to force any of the replacement options ('b' => blank lines,
'c' => comments including the original source lintes, and 'e' => empty
comment lines).

gcc/ada/ChangeLog:

	* doc/gnat_ugn/building_executable_programs_with_gnat.rst: Add doc
	for addition of -gnateG options b, c, and e.
	* doc/gnat_ugn/the_gnat_compilation_model.rst: Add doc for "-e" optionn
	on gnatprep and integrated preprocessing. Add doc for addition of
	-gnateG options b, c, and e.
	* gprep.adb (Scan_Command_Line): Add 'e' to the list of allowed switch
	letters in the string passed to GNAT.Command_Line.Getopt. Set the flag
	Opt.Empty_Comment_Deleted_Lines to True when the "-e" switch is found.
	(Usage): Output a switch description for the "-e" switch.
	* opt.ads: Add new flag variable Empty_Comment_Deleted_Lines. Add
	"GNAT" to the "GNATPREP" comment line for Comment_Deleted_Lines.
	* prep.adb (Output_Line): Add handling for Empty_Comment_Deleted_Lines,
	outputting comment lines consisting of "--!" for lines that are removed
	in the preprocessed source file when Empty_Comment_Deleted_Lines is True.
	* prepcomp.adb (Preproc_Data): Add Empty_Comments component (defaulting
	to False).
	(No_Preproc_Data): Add association for Empty_Comments component.
	(Parse_Preprocessing_Data_File): Add handling for new switch -e.
	(Prepare_To_Preprocess): Add logic for setting the new option
	Empty_Comment_Deleted_Lines (and making it the default for
	integrated preprocessing in the absence of other switches).
	* switch-c.adb (Scan_Front_End_Switches): Add support adding a single
	character 'b', 'c', 'e' to the "-gnateG" switch, to select any of the
	three options for replacing deleted lines in preprocessing output file.
	* usage.adb: Update usage info for -gnateG, to reflect the option of
	appending b, c, or e to the switch.
	* gnat_ugn.texi: Regenerate.