libstdc++: Rework handling of ISO week calendar and week index formatting.
The handling of ISO week-calendar year specifiers (%G, %g) and ISO week
number (%V) was merged into a single _M_g_G_V function, as the latter
requires ISO year value, computed by the former.

The values for %U and %W, which are based on the number of days since the
first Sunday and Monday of the year respectively, are now expressed as an
offset from the existing _M_day_of_year field. This reduces redundant
computation. The required flags were also updated to only need _DayOfYear
and _Weekday.

The _M_g_G_V function uses _M_day_of_year to compute __idoy, the day of the
year for the nearest Thursday. This value is used to determine if the ISO
year is the previous year (__idoy <= 0), the current year (__idoy <= 365/366),
next year (__idoy <= 730), or later year. This avoids an expensive conversion
from local_days to year_month_day in most cases. If the ISO calendar year
is current year, the __idoy value is reused for weekday index computation.

libstdc++-v3/ChangeLog:

	* include/bits/chrono_io.h(__formatter_chrono::_M_parse): Update
	needed flags for %g, %G, %V, %U, %W.
	(__formatter_chrono::_M_format_to): Change how %V is handled.
	(__formatter_chrono::_M_g_G): Merged into _M_g_G_V.
	(__formatter_chrono::_M_g_G_V): Reworked from _M_g_G.
	(__formatter_chrono::_M_U_V_W): Changed into _M_U_V.
	(__formatter_chrono::_M_U_W): Reworked implementation.
	* testsuite/std/time/year_month_day/io.cc: New tests.

Reviewed-by: Jonathan Wakely <jwakely@redhat.com>
Signed-off-by: Tomasz Kamiński <tkaminsk@redhat.com>